requirements.txt
aiogram==2.9.2
cchardet==2.1.6
aiodns==2.0.0
aiohttp==3.6.2
matplotlib==3.2.2
Pillow==7.1.2
numpy==1.19.0
torch==1.5.1
torchvision==0.6.1